Solution #4: Better Worker Load Balancing
The core problem here is that gevent is invalidating uwsgi's load balancing
algorithms assumption that the first worker that will accept the request is
the best worker to send the request to.

To prevent queueing we can terminate HTTP with nginx and proxy back to uwsgi.
This is pretty standard anyways because nginx is a really good app server,
but doing this can seriously help your performance on mixed workloads as well

Demo:

ab -c 4 -n 100 'http://127.0.0.1:1234/mixed/0.2'

Show latency heatmap showing just the perf we expect and top/uwsgitop output
with equal work distribution